+++ src/display/stellar/display.hpp
@@ -6,7 +6,7 @@
 namespace display {
     const int WIDTH = 16;
     const int HEIGHT = 16;
-    const int BUFFER_SIZE = WIDTH * HEIGHT * 4;
+    const size_t BUFFER_SIZE = WIDTH * HEIGHT * 4;
 
     void init();
     void update();
